+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/exceptions/UC4Exception.scala
@@ -1,6 +1,7 @@
 package de.upb.cs.uc4.shared.client.exceptions
 
 import com.lightbend.lagom.scaladsl.api.transport.TransportErrorCode
+import de.upb.cs.uc4.shared.client.exceptions.ErrorType.ErrorType
 
 abstract class UC4Exception(
     val errorCode: TransportErrorCode,
@@ -24,6 +25,8 @@ object UC4Exception {
 
   //400
   val DeserializationError = new UC4NonCriticalException(400, GenericError(ErrorType.Deserialization))
+  def KafkaDeserializationError(expected: String, actual: String) = new UC4NonCriticalException(400, GenericError(ErrorType.KafkaDeserialization, s"Expected class type '$expected' but received '$actual''"))
+
   val MalformedRefreshToken = new UC4NonCriticalException(400, GenericError(ErrorType.MalformedRefreshToken))
   val MalformedLoginToken = new UC4NonCriticalException(400, GenericError(ErrorType.MalformedLoginToken))
   val MultipleAuthorizationError = new UC4NonCriticalException(400, GenericError(ErrorType.MultipleAuthorization))