+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/kafka/EncryptionContainer.scala
@@ -1,0 +1,14 @@
+package de.upb.cs.uc4.shared.client.kafka
+
+import play.api.libs.json.{ Format, Json }
+
+/** Wraps an encrypted object and its original class type
+  *
+  * @param classType of the encrypted object, to be consistent use the canonical name of the class
+  * @param data encrypted payload object
+  */
+case class EncryptionContainer(classType: String, data: Array[Byte])
+
+object EncryptionContainer {
+  implicit val format: Format[EncryptionContainer] = Json.format
+}