+++ product_code/user_service/api/src/test/scala/de/upb/cs/uc4/user/UserServiceStub.scala
@@ -6,6 +6,7 @@ import com.lightbend.lagom.scaladsl.api.ServiceCall
 import com.lightbend.lagom.scaladsl.api.broker.Topic
 import de.upb.cs.uc4.authentication.model.JsonUsername
 import de.upb.cs.uc4.shared.client.exceptions.UC4Exception
+import de.upb.cs.uc4.shared.client.kafka.EncryptionContainer
 import de.upb.cs.uc4.user.api.UserService
 import de.upb.cs.uc4.user.model._
 import de.upb.cs.uc4.user.model.post.PostMessageUser
@@ -88,9 +89,9 @@ class UserServiceStub extends UserService with DefaultTestUsers {
 
   override def allowedDeleteGetPut: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }
 
-  override def userCreationTopic(): Topic[Usernames] = null
+  override def userCreationTopic(): Topic[EncryptionContainer] = null //EncryptionContainer[Usernames]
 
-  override def userDeletedTopic(): Topic[JsonUsername] = null
+  override def userDeletionTopic(): Topic[EncryptionContainer] = null //EncryptionContainer[JsonUsername]
 
   override def allowVersionNumber: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }